+++ module/schemas/commonSchemas.js
@@ -21,6 +21,23 @@ export class NullableEmbeddedDataField extends fields.EmbeddedDataField {
   }
 }
 
+export class OptionalSchemaField extends fields.SchemaField {
+  /**
+   * @param {typeof DataModel} model          The class of DataModel which should be embedded in this field
+   * @param {DataFieldOptions} options        Options which configure the behavior of the field
+   */
+  constructor(fields, options = {}) {
+    super(fields, options);
+  }
+
+  toObject(value) {
+    if (value != undefined) {
+      return super.toObject(value);
+    }
+    return undefined;
+  }
+}
+
 export class NullableDocumentIdField extends fields.DocumentIdField {
   /** @inheritdoc */
   static get _defaults() {