+++ module/tests/rollsTest.js
@@ -428,7 +428,7 @@ export function registerRollTesting(quench) {
               magus.system.arts.techniques.mu.finalScore +
               magus.system.arts.forms.co.finalScore +
               magus.system.characteristics.sta.value +
-              magus.system.woundsTotal +
+              magus.system.penalties.wounds.total +
               magus.system.fatigueTotal +
               3;
             assert.equal(roll.modifier(), tot);
@@ -460,7 +460,7 @@ export function registerRollTesting(quench) {
               magus.system.arts.techniques.pe.finalScore +
               magus.system.arts.forms.co.finalScore +
               magus.system.characteristics.sta.value +
-              magus.system.woundsTotal +
+              magus.system.penalties.wounds.total +
               magus.system.fatigueTotal +
               1 +
               3;
@@ -491,7 +491,7 @@ export function registerRollTesting(quench) {
               magus.system.arts.techniques.cr.finalScore +
               magus.system.arts.forms.ig.finalScore +
               magus.system.characteristics.sta.value +
-              magus.system.woundsTotal +
+              magus.system.penalties.wounds.total +
               magus.system.fatigueTotal +
               2 +
               3;
@@ -523,7 +523,7 @@ export function registerRollTesting(quench) {
               // magus.system.arts.techniques.cr.finalScore +
               // magus.system.arts.forms.ig.finalScore +
               magus.system.characteristics.sta.value +
-              magus.system.woundsTotal +
+              magus.system.penalties.wounds.total +
               magus.system.fatigueTotal +
               3;
             assert.equal(roll.modifier(), tot);
@@ -556,7 +556,7 @@ export function registerRollTesting(quench) {
               magus.system.arts.techniques.mu.finalScore +
               magus.system.arts.forms.an.finalScore * 2 +
               magus.system.characteristics.sta.value +
-              magus.system.woundsTotal +
+              magus.system.penalties.wounds.total +
               magus.system.fatigueTotal +
               3;
             assert.equal(roll.modifier(), tot);
@@ -589,7 +589,7 @@ export function registerRollTesting(quench) {
               magus.system.arts.techniques.mu.finalScore +
               magus.system.arts.forms.im.finalScore +
               magus.system.characteristics.sta.value +
-              magus.system.woundsTotal +
+              magus.system.penalties.wounds.total +
               magus.system.fatigueTotal +
               Sp1.system.mastery +
               Sp1.system.bonus +
@@ -624,7 +624,7 @@ export function registerRollTesting(quench) {
               magus.system.arts.techniques.mu.finalScore +
               magus.system.arts.forms.co.finalScore * 2 +
               magus.system.characteristics.sta.value +
-              magus.system.woundsTotal +
+              magus.system.penalties.wounds.total +
               magus.system.fatigueTotal +
               Sp2.system.mastery +
               Sp2.system.bonus +
@@ -663,7 +663,7 @@ export function registerRollTesting(quench) {
               magus.system.arts.techniques.mu.finalScore +
               magus.system.arts.forms.co.finalScore * 2 +
               magus.system.characteristics.sta.value +
-              magus.system.woundsTotal +
+              magus.system.penalties.wounds.total +
               magus.system.fatigueTotal +
               Sp2.system.mastery +
               Sp2.system.bonus +
@@ -698,7 +698,7 @@ export function registerRollTesting(quench) {
               magus.system.arts.techniques.mu.finalScore +
               magus.system.arts.forms.vi.finalScore +
               magus.system.characteristics.sta.value +
-              magus.system.woundsTotal +
+              magus.system.penalties.wounds.total +
               magus.system.fatigueTotal +
               Sp3.system.mastery +
               Sp3.system.bonus +